Profile details, скрытие аватарки не отрабатывает Начал работать над карточкой просмотра поста: Загрузка карт работает

+++ app/src/main/java/ru/kostenko/nework/dto/Coords.kt
@@ -1,7 +1,7 @@
 package ru.kostenko.nework.dto
 
 data class Coords(
-    val latC: Int = 0,
-    val longC: Int = 0,
+    val latC: Double = 0.0,
+    val longC: Double = 0.0,
     )
