+++ app/src/main/java/ru/kostenko/nework/entity/CoordsEntity.kt
@@ -1,11 +1,11 @@
 package ru.kostenko.nework.entity
 
-import androidx.room.Entity
+
 import ru.kostenko.nework.dto.Coords
 
 data class CoordsEntity(
-    val coordslat: Int = 0,
-    val coordslong: Int = 0,
+    val coordslat: Double = 0.0,
+    val coordslong: Double = 0.0,
 ) {
     fun toDto() = Coords(coordslat, coordslong)
